logic/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
psycopg2==2.9.9
pydantic==2.5.3
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
numpy==1.26.3
PyJWT==2.8.0
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1
alembic==1.13.0
redis==5.0.1
pytest==7.4.3
pytest-asyncio==0.23.2
black==23.11.0
mypy==1.7.1
ruff==0.1.6 